src/python/requirements.txt
# Web Framework
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7

# Database
sqlalchemy==2.0.20
psycopg2==2.9.9  # For PostgreSQL support

# Environment and Configuration
python-dotenv==1.0.0

# Data Processing and Analysis
nltk==3.8.1
textblob==0.17.1
pandas==2.2.3
numpy==1.24.3

# Task Scheduling
apscheduler==3.10.1

# Social Media APIs
tweepy==4.14.0  # Twitter
praw==7.7.1     # Reddit
facebook-sdk==3.1.0  # Facebook
instaloader==4.10.0   # Instagram
linkedin-api==2.0.0a5  # LinkedIn

# HTTP and API
requests==2.31.0

# Development Tools
pytest==7.1.2  # For testing
black==22.3.0  # For code formatting
flake8==4.0.1  # For linting

# Additional dependencies
beautifulsoup4==4.12.2
selenium==4.11.2
webdriver-manager==4.0.0